requirements.txt
python-dotenv
flask==3.1.0
anthropic==0.40.0
werkzeug==3.1.3
gunicorn==23.0.0
psycopg[binary]==3.2.3
pytest